USER STORY 1:DEPENDENCY INJECTION
Create a class MyCategory that implements CategoryInterface, set the
preference in di.xml of CategoryInterface to MyCategory.

Create another class Test, inject CategoryInterface here in the
constructor, also pass string and array as param.

Create a controller, and inject the Test class into it.
In order for pref to work, the class implementing the interface MUST
implement ALL the methods of interface.